requirements.txt
Flask
Flask-SQLAlchemy
bcrypt
flask_login
playwright